ADD LeadersBoard Implementaion and the API call for retieval/storage
In this commit, I implemented the scoring board dispay functionality. I also used the leardersboard api service to get and store scores for the game. I also implemented localStorage and added a bit of styling to the game as well. I updated the README documentation to add onto the already accoplished objectives. In it, I added screnshots of the game to act as a guide on what the game is all about.